fix(ai-chat): show new conversations in Recents on chat page
Extract useConversationListSync into a shared hook (with a shared
ChatRuntimeContext) and wire it into AiChatPage so the page gets the
same optimistic insert and post-stream refetch behavior the drawer
already had. Previously the page's RecentConversations only refreshed
on incidental refetches.
