+++ plugins/ai-chat/src/components/AiChatDrawer/AiChatDrawerProvider.tsx
@@ -1,18 +1,12 @@
 import {
   ReactNode,
-  createContext,
   useCallback,
-  useContext,
   useEffect,
   useMemo,
   useRef,
   useState,
 } from 'react';
-import {
-  AssistantRuntimeProvider,
-  useAssistantApi,
-  useAssistantState,
-} from '@assistant-ui/react';
+import { AssistantRuntimeProvider, useAssistantApi } from '@assistant-ui/react';
 import {
   useApi,
   discoveryApiRef,
@@ -22,28 +16,16 @@ import { useMediaQuery, useTheme } from '@material-ui/core';
 import { aiChatDrawerApiRef } from '@giantswarm/backstage-plugin-ai-chat-react';
 import { UIMessage } from 'ai';
 import { useChatSetup, UseChatSetupOptions } from '../../hooks/useChatSetup';
-import { useConversations } from '../../hooks/useConversations';
 import {
-  ConversationClient,
-  ConversationApi,
-  ConversationListItem,
-} from '../../api';
+  ChatRuntimeContext,
+  useChatRuntimeContext,
+} from '../../hooks/ChatRuntimeContext';
+import { useConversationListSync } from '../../hooks/useConversationListSync';
+import { ConversationClient, ConversationApi } from '../../api';
 import { AiChatDrawer, AiChatDrawerVariant } from './AiChatDrawer';
 import { QueryClientProvider } from '../QueryClientProvider';
 
 export type DrawerTab = 'chat' | 'history';
-
-interface RuntimeContextValue {
-  isReady: boolean;
-  getConversationId: () => string | null;
-  isNewConversation: boolean;
-}
-
-const RuntimeContext = createContext<RuntimeContextValue>({
-  isReady: false,
-  getConversationId: () => null,
-  isNewConversation: true,
-});
 
 /**
  * Owns the chat runtime. Keyed externally so that remounting creates a
@@ -72,91 +54,11 @@ const AiChatRuntimeProvider = ({
   );
   return (
     <AssistantRuntimeProvider runtime={runtime}>
-      <RuntimeContext.Provider value={value}>
+      <ChatRuntimeContext.Provider value={value}>
         {children}
-      </RuntimeContext.Provider>
+      </ChatRuntimeContext.Provider>
     </AssistantRuntimeProvider>
   );
-};
-
-type ThreadMessage = {
-  readonly role: string;
-  readonly content: ReadonlyArray<{
-    readonly type: string;
-    readonly text?: string;
-  }>;
-};
-
-function extractFirstUserText(messages: ReadonlyArray<ThreadMessage>): string {
-  const firstUser = messages.find(m => m.role === 'user');
-  if (!firstUser) return '';
-  return firstUser.content
-    .filter(p => p.type === 'text' && typeof p.text === 'string')
-    .map(p => p.text as string)
-    .join(' ')
-    .trim();
-}
-
-/**
- * Keeps the conversation history list in sync with chat activity:
- * - On the first user message of a brand-new conversation, optimistically
- *   prepends a synthetic ConversationListItem so the entry shows up in the
- *   History tab immediately, before the backend has persisted the row.
- * - When the assistant stream finishes (isRunning transitions to false),
- *   invalidates the list query so the cached entry reconciles against the
- *   server's authoritative copy (real preview, real updatedAt).
- */
-const useConversationListSync = (conversationApi: ConversationApi) => {
-  const { getConversationId, isNewConversation } = useContext(RuntimeContext);
-  const { addOptimisticConversation, refreshConversations } =
-    useConversations(conversationApi);
-  const insertedRef = useRef(false);
-  const wasRunningRef = useRef(false);
-
-  const isRunning = useAssistantState(({ thread }) =>
-    Boolean(thread?.isRunning),
-  );
-  const messages = useAssistantState(
-    ({ thread }) =>
-      thread?.messages as ReadonlyArray<ThreadMessage> | undefined,
-  );
-
-  // Optimistic insert on the first user message of a new conversation.
-  useEffect(() => {
-    if (insertedRef.current) return;
-    if (!isNewConversation) return;
-    if (!messages || messages.length === 0) return;
-    const firstUserText = extractFirstUserText(messages);
-    if (!firstUserText) return;
-    const id = getConversationId();
-    if (!id) return;
-
-    const now = new Date().toISOString();
-    const optimistic: ConversationListItem = {
-      id,
-      userId: '',
-      title: undefined,
-      preview: firstUserText,
-      isStarred: false,
-      createdAt: now,
-      updatedAt: now,
-    };
-    addOptimisticConversation(optimistic);
-    insertedRef.current = true;
-  }, [
-    messages,
-    isNewConversation,
-    getConversationId,
-    addOptimisticConversation,
-  ]);
-
-  // Reconcile with the server when the stream finishes.
-  useEffect(() => {
-    if (wasRunningRef.current && !isRunning) {
-      refreshConversations();
-    }
-    wasRunningRef.current = isRunning;
-  }, [isRunning, refreshConversations]);
 };
 
 /**
@@ -185,7 +87,7 @@ const DrawerInner = ({
   onSelectConversation(id: string): void;
 }) => {
   const assistantApi = useAssistantApi();
-  const { isReady } = useContext(RuntimeContext);
+  const { isReady } = useChatRuntimeContext();
 
   useConversationListSync(conversationApi);
